正規表現マッチメソッドをHelperに分離して単体テストを追加
