+++ spotter/Classes/Controllers/TimeLineViewController.swift
@@ -49,7 +49,7 @@ class TimeLineViewController: UIViewController, UITableViewDelegate, UITableView
         cell.faceImageView.kf.setImage(with: profile_url)
         
         let tweetText = microposts[indexPath.row].content
-        let urlTextArray = getMatchStrings(targetString: tweetText, pattern: "(http://|https://){1}[\\w\\.\\-/:]+")
+        let urlTextArray = TimelineHelper.getMatchStrings(targetString: tweetText, pattern: "(http://|https://){1}[\\w\\.\\-/:]+")
         if (urlTextArray.count == 0) {
             cell.tweetText.text = tweetText
             return cell
@@ -64,7 +64,7 @@ class TimeLineViewController: UIViewController, UITableViewDelegate, UITableView
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let cell = timelineView.cellForRow(at: indexPath) as! TimelineTableViewCell
-        let urlTextArray = getMatchStrings(targetString: cell.tweetText.text!, pattern: "(http://|https://){1}[\\w\\.\\-/:]+")
+        let urlTextArray = TimelineHelper.getMatchStrings(targetString: cell.tweetText.text!, pattern: "(http://|https://){1}[\\w\\.\\-/:]+")
         if (urlTextArray.count != 0) {
             let url = URL(string: urlTextArray[0])
             if UIApplication.shared.canOpenURL(url!) {
@@ -100,25 +100,5 @@ class TimeLineViewController: UIViewController, UITableViewDelegate, UITableView
         DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
             sender.endRefreshing()
         }
-    }
-    
-    // 正規表現にマッチした文字列を格納した配列を返す
-    func getMatchStrings(targetString: String, pattern: String) -> [String] {
-        var matchStrings:[String] = []
-        do {
-            let regex = try NSRegularExpression(pattern: pattern, options: [])
-            let targetStringRange = NSRange(location: 0, length: (targetString as NSString).length)
-            
-            let matches = regex.matches(in: targetString, options: [], range: targetStringRange)
-            
-            for match in matches {
-                let range = match.range(at: 0)
-                let result = (targetString as NSString).substring(with: range)
-                
-                matchStrings.append(result)
-            }
-            return matchStrings
-        } catch {}
-        return []
     }
 }
